requirements.txt
-e .
aws_cdk.core
aws_cdk.aws_s3
aws_cdk.aws_ec2